created function dbBroadcast in MessageHandler
