+++ be2-scala/src/main/scala/ch/epfl/pop/pubsub/graph/handlers/MessageHandler.scala
@@ -3,13 +3,14 @@ package ch.epfl.pop.pubsub.graph.handlers
 import akka.pattern.AskableActorRef
 import ch.epfl.pop.model.network.JsonRpcRequest
 import ch.epfl.pop.model.network.method.message.Message
+import ch.epfl.pop.model.objects.{Base64Data, Channel, DbActorNAckException, Hash, Signature}
 import ch.epfl.pop.pubsub.AskPatternConstants
 import ch.epfl.pop.pubsub.graph.{ErrorCodes, GraphMessage, PipelineError}
 import ch.epfl.pop.storage.DbActor
 
 import scala.concurrent.ExecutionContext.Implicits.global
-import scala.concurrent.Future
-import scala.util.Success
+import scala.concurrent.{Await, Future}
+import scala.util.{Failure, Success}
 
 trait MessageHandler extends AskPatternConstants {
 
@@ -56,6 +57,35 @@ trait MessageHandler extends AskPatternConstants {
     askWritePropagate.transformWith {
       case Success(_) => Future(Left(rpcRequest))
       case _ => Future(Right(PipelineError(ErrorCodes.SERVER_ERROR.id, s"dbAskWritePropagate failed : could not write & propagate message $m", rpcRequest.id)))
+    }
+  }
+
+  /**
+   * Helper function for broadcasting messages to subscribers
+   *
+   * @param rpcMessage       : message for which we want to generate the broadcast
+   * @param channel          : the Channel in which we read the data
+   * @param broadcastData    : the message data we broadcast converted to Base64Data
+   * @param broadcastChannel : the Channel in which we broadcast
+   */
+  def dbBroadcast(rpcMessage: JsonRpcRequest, channel: Channel, broadcastData: Base64Data, broadcastChannel: Channel): GraphMessage = {
+    val askLaoData = dbActor ? DbActor.ReadLaoData(channel)
+
+    Await.ready(askLaoData, duration).value match {
+      case Some(Success(DbActor.DbActorReadLaoDataAck(laoData))) =>
+        val broadcastSignature: Signature = laoData.privateKey.signData(broadcastData)
+        val broadcastId: Hash = Hash.fromStrings(broadcastData.toString, broadcastSignature.toString)
+        val broadcastMessage: Message = Message(broadcastData, laoData.publicKey, broadcastSignature, broadcastId, List.empty)
+
+        val askWritePropagate = dbActor ? DbActor.WriteAndPropagate(broadcastChannel, broadcastMessage)
+        Await.ready(askWritePropagate, duration).value.get match {
+          case Success(_) => Left(rpcMessage)
+          case Failure(ex: DbActorNAckException) => Right(PipelineError(ex.code, s"broadcastHelper failed : ${ex.message}", rpcMessage.getId))
+          case reply => Right(PipelineError(ErrorCodes.SERVER_ERROR.id, s"broadcastHelper failed : unknown DbActor reply $reply", rpcMessage.getId))
+        }
+
+      case Some(Failure(ex: DbActorNAckException)) => Right(PipelineError(ex.code, s"broadcastHelper failed : ${ex.message}", rpcMessage.getId))
+      case reply => Right(PipelineError(ErrorCodes.SERVER_ERROR.id, s"broadcastHelper failed : unknown DbActor reply $reply", rpcMessage.getId))
     }
   }
 }
